+++ mountpoint-s3/tests/fuse_tests/read_test.rs
@@ -1,5 +1,5 @@
 use std::fs::{read_dir, File};
-use std::io::{Read as _, Seek, SeekFrom, Write};
+use std::io::{Read, Seek, SeekFrom, Write};
 #[cfg(not(feature = "s3express_tests"))]
 use std::os::unix::prelude::PermissionsExt;
 use std::path::Path;
@@ -273,10 +273,11 @@ where
     assert_eq!(dir_entry_names, vec!["hello.txt"]);
 
     // Overwrite the test file and verify that we can't read from a file opened in O_WRONLY
-    let mut write_fh = File::options().write(true).open(&file_path).unwrap();
+    let mut write_fh = File::options().write(true).truncate(true).open(&file_path).unwrap();
     let mut contents = String::new();
     let err = write_fh.read_to_string(&mut contents).expect_err("read should fail");
     assert_eq!(err.raw_os_error(), Some(libc::EBADF));
+    write_fh.sync_all().unwrap();
     drop(write_fh);
 
     // We shouldn't be able to read from a file mid-write in O_RDWR
@@ -293,8 +294,8 @@ where
     assert_eq!(err.raw_os_error(), Some(libc::EBADF));
 
     // Read should also fail from different file handle
-    let mut read_fh = open_for_read(&file_path, true).unwrap();
-    let err = read_fh.read_to_string(&mut contents).expect_err("read should fail");
+    let err =
+        open_for_read(&file_path, true).expect_err("opening for read should fail with pending write handles open");
     assert_eq!(err.raw_os_error(), Some(libc::EPERM));
 }
 
@@ -308,4 +309,45 @@ fn read_errors_test_s3() {
 #[test_case("read_errors_test"; "prefix")]
 fn read_errors_test_mock(prefix: &str) {
     read_errors_test(fuse::mock_session::new, prefix);
+}
+
+fn read_after_flush_test<F>(creator_fn: F)
+where
+    F: FnOnce(&str, TestSessionConfig) -> (TempDir, BackgroundSession, TestClientBox),
+{
+    const KEY: &str = "data.bin";
+    let (mount_point, _session, mut test_client) = creator_fn("read_after_flush_test", Default::default());
+
+    let mut rng = ChaChaRng::seed_from_u64(0x87654321);
+    let mut two_mib_body = vec![0; 2 * 1024 * 1024];
+    rng.fill_bytes(&mut two_mib_body);
+    test_client.put_object(KEY, &two_mib_body).unwrap();
+
+    let path = mount_point.path().join(KEY);
+    let mut f = open_for_read(path, true).unwrap();
+
+    let mut content = vec![0; 128];
+    f.read_exact(&mut content).unwrap();
+
+    let mut f_dup = f.try_clone().unwrap();
+
+    // Close the file. Triggers a flush on the file handle.
+    drop(f);
+
+    // Read using the duplicated instance (same underlying handle).
+    // Seek to the end of the file to avoid relying on the kernel cache.
+    let pos = f_dup.seek(SeekFrom::End(-(content.len() as i64))).unwrap() as usize;
+    f_dup.read_exact(&mut content).unwrap();
+    assert_eq!(content, two_mib_body[pos..]);
+}
+
+#[cfg(feature = "s3_tests")]
+#[test]
+fn read_after_flush_test_s3() {
+    read_after_flush_test(fuse::s3_session::new);
+}
+
+#[test]
+fn read_after_flush_test_mock() {
+    read_after_flush_test(fuse::mock_session::new);
 }